style(reporters): move to file-scoped namespace
