+++ src/Stryker.Core/Stryker.Core.UnitTest/Reporters/ReporterFactoryTests.cs
@@ -5,7 +5,7 @@ using Shouldly;
 using Stryker.Core.Baseline.Providers;
 using Stryker.Core.Options;
 using Stryker.Core.Reporters;
-using Stryker.Core.Reporters.Html.reporter;
+using Stryker.Core.Reporters.HtmlReporter;
 using Stryker.Core.Reporters.Json;
 using Stryker.Core.Reporters.Progress;
 using Xunit;
